notification-microservice/requirements.txt
Flask==2.0.2
pika==1.2.0
Werkzeug==2.2.2
python-dotenv
requests 